requirements.txt
fastapi
uvicorn
boto3
python-dotenv
databases
aiosqlite
aws-cdk-lib==2.166.0